Se implementan las pruebas de integración para el registro de ingreso de carros

+++ estacionamiento/src/main/java/co/ceiba/adn/estacionamiento/entity/Motorcycle.java
@@ -20,7 +20,7 @@ public class Motorcycle extends Vehicle {
 		this.cylinderCapacity = cylinderCapacity;
 	}
 
-	public int getCylinderCapacity() {
+	public short getCylinderCapacity() {
 		return cylinderCapacity;
 	}
